+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v1.29.0
 
   - Don't make `Kernel#warn` redirection to `Console.warn` the default behavior, you must `require 'console/warn'` to enable it.
   - Remove deprecated `Console::Logger#failure`.
@@ -9,7 +9,7 @@
 
 `Console.call` and all wrapper methods will now consistently handle exceptions that are the last positional argument or keyword argument. This means that the following code will work as expected:
 
-```ruby
+``` ruby
 begin
 rescue => error
 	# Last positional argument: